+++ apps/io-session-manager-oi/CHANGELOG.md
@@ -1,5 +1,18 @@
 # io-session-manager-oi
 
+## 0.18.0
+
+### Minor Changes
+
+- 55db1f1: implement getSession handler and use case and mount it to fastify server
+
+### Patch Changes
+
+- Updated dependencies [55db1f1]
+- Updated dependencies [55db1f1]
+  - @pagopa/io-auth-n-identity-session@0.4.0
+  - @pagopa/io-auth-n-identity-domain@0.5.1
+
 ## 0.17.0
 
 ### Minor Changes